feat(cli)!: split shep import into pm2 and env subcommands
Bare `shep import` no longer parses; the pm2 dump reader moves to
`shep import pm2` with its four flags (--from, --out, --dry-run,
--force) unchanged. commands::import::import becomes a dispatch
function over ImportCommand; commands::import::pm2::import holds
today's logic verbatim. `env` arrives in a later commit.
